requirements.txt
# Core scientific stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration
pyyaml>=6.0

# Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0

# Optional: statsmodels for FDR/BH correction
statsmodels>=0.14.0

# Notebook support (optional, not required for CI)
# jupyter>=1.0.0
